+++ src/fonduer/supervision/labeler.py
@@ -306,9 +306,14 @@ class Labeler(UDFRunner):
         # Clear Labels for the candidates in the split passed in.
         logger.info(f"Clearing Labels (split {split})")
 
-        sub_query = (
-            self.session.query(Candidate.id).filter(Candidate.split == split).subquery()
-        )
+        if split == ALL_SPLITS:
+            sub_query = self.session.query(Candidate.id).subquery()
+        else:
+            sub_query = (
+                self.session.query(Candidate.id)
+                .filter(Candidate.split == split)
+                .subquery()
+            )
         query = self.session.query(Label).filter(Label.candidate_id.in_(sub_query))
         query.delete(synchronize_session="fetch")