+++ CHANGELOG.rst
@@ -101,6 +101,8 @@ Fixed
 * `@HiromuHota`_: Fix deadlock error during Labeler.apply and Featurizer.apply.
   (`#328 <https://github.com/HazyResearch/fonduer/issues/328>`_)
 * `@HiromuHota`_: Avoid networkx 2.4 so that snorkel-metal does not use the removed API.
+* `@HiromuHota`_: Fix the issue that Labeler.apply with docs instead of split fails.
+  (`#340 <https://github.com/HazyResearch/fonduer/pull/340>`_)
 
 0.7.0_ - 2019-06-12
 -------------------